+++ lotti/lib/widgets/create/add_tag_actions.dart
@@ -1,8 +1,7 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
-import 'package:lotti/classes/tag_type_definitions.dart';
-import 'package:lotti/pages/settings/tags/tag_details.dart';
+import 'package:lotti/routes/router.gr.dart';
 import 'package:lotti/theme.dart';
-import 'package:lotti/utils/file_utils.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:radial_button/widget/circle_floating_button.dart';
 
@@ -38,21 +37,9 @@ class _RadialAddTagButtonsState extends State<RadialAddTagButtons> {
         ),
         backgroundColor: AppColors.entryBgColor,
         onPressed: () {
-          Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (BuildContext context) {
-                DateTime now = DateTime.now();
-                return TagDetailRoute(
-                  tagEntity: TagEntity.genericTag(
-                    id: uuid.v1(),
-                    vectorClock: null,
-                    createdAt: now,
-                    updatedAt: now,
-                    private: false,
-                    tag: '',
-                  ),
-                );
-              },
+          context.router.push(
+            CreateTagRoute(
+              tagType: 'TAG',
             ),
           );
         },
@@ -65,21 +52,9 @@ class _RadialAddTagButtonsState extends State<RadialAddTagButtons> {
         ),
         backgroundColor: AppColors.entryBgColor,
         onPressed: () {
-          Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (BuildContext context) {
-                DateTime now = DateTime.now();
-                return TagDetailRoute(
-                  tagEntity: TagEntity.personTag(
-                    id: uuid.v1(),
-                    vectorClock: null,
-                    createdAt: now,
-                    updatedAt: now,
-                    private: false,
-                    tag: '',
-                  ),
-                );
-              },
+          context.router.push(
+            CreateTagRoute(
+              tagType: 'PERSON',
             ),
           );
         },
@@ -92,21 +67,9 @@ class _RadialAddTagButtonsState extends State<RadialAddTagButtons> {
         ),
         backgroundColor: AppColors.entryBgColor,
         onPressed: () {
-          Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (BuildContext context) {
-                DateTime now = DateTime.now();
-                return TagDetailRoute(
-                  tagEntity: TagEntity.storyTag(
-                    id: uuid.v1(),
-                    vectorClock: null,
-                    createdAt: now,
-                    updatedAt: now,
-                    private: false,
-                    tag: '',
-                  ),
-                );
-              },
+          context.router.push(
+            CreateTagRoute(
+              tagType: 'STORY',
             ),
           );
         },
